added auth info
