+++ .gitignore
@@ -23,4 +23,5 @@ yarn-debug.log*
 yarn-error.log*
 
 package-lock.json
-
+src/helpers/apiKeys.json
+.firebase/
