+++ src/components/Home/Home.scss
@@ -1,0 +1,3 @@
+.Home {
+  text-align: center;
+}
