+++ src/helpers/data/authData.js
@@ -1,0 +1,7 @@
+import firebase from 'firebase/app';
+import 'firebase/auth';
+
+
+const getUid = () => firebase.auth().currentUser.uid;
+
+export default { getUid };
