+++ src/helpers/data/connections.js
@@ -1,0 +1,10 @@
+import firebase from 'firebase/app';
+import firebaseConfig from '../apiKeys.json';
+
+const firebaseApp = () => {
+  if (!firebase.apps.length) {
+    firebase.initializeApp(firebaseConfig.firebaseKeys);
+  }
+};
+
+export default firebaseApp;
